use deterministic censusID (#77)
use deterministic censusID

Based on blockNumber and StrategyID we can derive
a unique identifier for a censusID.
This way we can easily check if the census already
exist in the database.

Bonus: update github CI configuration

---------

+++ api/strategy.go
@@ -72,9 +72,9 @@ func (capi *census3API) getStrategies(msg *api.APIdata, ctx *httprouter.HTTPCont
 		return ErrNoStrategies
 	}
 	// parse and encode the strategies
-	strategies := GetStrategiesResponse{Strategies: []uint64{}}
+	strategies := GetStrategiesResponse{Strategies: []uint32{}}
 	for _, strategy := range rows {
-		strategies.Strategies = append(strategies.Strategies, uint64(strategy.ID))
+		strategies.Strategies = append(strategies.Strategies, uint32(strategy.ID))
 	}
 	res, err := json.Marshal(strategies)
 	if err != nil {
@@ -151,9 +151,9 @@ func (capi *census3API) getTokenStrategies(msg *api.APIdata, ctx *httprouter.HTT
 		return ErrNoStrategies
 	}
 	// parse and encode strategies
-	strategies := GetStrategiesResponse{Strategies: []uint64{}}
+	strategies := GetStrategiesResponse{Strategies: []uint32{}}
 	for _, tokenStrategy := range rows {
-		strategies.Strategies = append(strategies.Strategies, uint64(tokenStrategy.ID))
+		strategies.Strategies = append(strategies.Strategies, uint32(tokenStrategy.ID))
 	}
 	res, err := json.Marshal(strategies)
 	if err != nil {
